--- HospitalJobs-Mobile/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Fri Jun 20 00:03:30 GST 2025
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-8.11.1-bin.zip
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
