--- HospitalJobs-Mobile/settings.gradle.kts
@@ -1,19 +1,0 @@
-enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
-pluginManagement {
-    repositories {
-        google()
-        gradlePluginPortal()
-        mavenCentral()
-    }
-}
-
-dependencyResolutionManagement {
-    repositories {
-        google()
-        mavenCentral()
-    }
-}
-
-rootProject.name = "HospitalJobs"
-include(":hospitaljobsin")
-include(":shared")
