--- HospitalJobs-Mobile/shared/src/androidMain/kotlin/com/nvadi/hospitaljobs/Platform.android.kt
@@ -1,7 +1,0 @@
-package com.nvadi.hospitaljobs
-
-class AndroidPlatform : Platform {
-    override val name: String = "Android ${android.os.Build.VERSION.SDK_INT}"
-}
-
-actual fun getPlatform(): Platform = AndroidPlatform()
