--- HospitalJobs-Mobile/shared/src/commonMain/kotlin/com/nvadi/hospitaljobs/Greeting.kt
@@ -1,9 +1,0 @@
-package com.nvadi.hospitaljobs
-
-class Greeting {
-    private val platform: Platform = getPlatform()
-
-    fun greet(): String {
-        return "Hello, ${platform.name}!"
-    }
-}
