--- HospitalJobs-Mobile/shared/src/commonMain/kotlin/com/nvadi/hospitaljobs/Platform.kt
@@ -1,7 +1,0 @@
-package com.nvadi.hospitaljobs
-
-interface Platform {
-    val name: String
-}
-
-expect fun getPlatform(): Platform
