--- HospitalJobs-Mobile/shared/src/iosMain/kotlin/com/nvadi/hospitaljobs/Platform.ios.kt
@@ -1,9 +1,0 @@
-package com.nvadi.hospitaljobs
-
-import platform.UIKit.UIDevice
-
-class IOSPlatform: Platform {
-    override val name: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
-}
-
-actual fun getPlatform(): Platform = IOSPlatform()
