+++ hospitaljobsin-Mobile/HospitalJobsIn/src/main/AndroidManifest.xml
