+++ hospitaljobsin-Mobile/HospitalJobsIn/src/main/res/values/styles.xml
