+++ hospitaljobsin-Mobile/gradle.properties
