+++ hospitaljobsin-Mobile/settings.gradle.kts
@@ -1,0 +1,19 @@
+enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
+pluginManagement {
+    repositories {
+        google()
+        gradlePluginPortal()
+        mavenCentral()
+    }
+}
+
+dependencyResolutionManagement {
+    repositories {
+        google()
+        mavenCentral()
+    }
+}
+
+rootProject.name = "HospitalJobsIn"
+include(":HospitalJobsIn")
+include(":shared")
