+++ hospitaljobsin-Mobile/shared/src/androidMain/kotlin/com/hospitaljobsin/hospitaljobsin/Platform.android.kt
@@ -1,0 +1,7 @@
+package com.hospitaljobsin.hospitaljobsin
+
+class AndroidPlatform : Platform {
+    override val name: String = "Android ${android.os.Build.VERSION.SDK_INT}"
+}
+
+actual fun getPlatform(): Platform = AndroidPlatform()
