+++ hospitaljobsin-Mobile/shared/src/commonMain/kotlin/com/hospitaljobsin/hospitaljobsin/Greeting.kt
@@ -1,0 +1,9 @@
+package com.hospitaljobsin.hospitaljobsin
+
+class Greeting {
+    private val platform: Platform = getPlatform()
+
+    fun greet(): String {
+        return "Hello, ${platform.name}!"
+    }
+}
