+++ hospitaljobsin-Mobile/shared/src/commonMain/kotlin/com/hospitaljobsin/hospitaljobsin/Platform.kt
@@ -1,0 +1,7 @@
+package com.hospitaljobsin.hospitaljobsin
+
+interface Platform {
+    val name: String
+}
+
+expect fun getPlatform(): Platform
