+++ hospitaljobsin-Mobile/shared/src/iosMain/kotlin/com/hospitaljobsin/hospitaljobsin/Platform.ios.kt
@@ -1,0 +1,9 @@
+package com.hospitaljobsin.hospitaljobsin
+
+import platform.UIKit.UIDevice
+
+class IOSPlatform: Platform {
+    override val name: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
+}
+
+actual fun getPlatform(): Platform = IOSPlatform()
